+++ scripts/data/building/building_catalog.gd
@@ -1,3 +1,4 @@
+@tool
 class_name BuildingCatalog
 extends Resource