+++ scripts/data/resource_nodes/resource_node_catalog.gd
@@ -1,3 +1,4 @@
+@tool
 class_name ResourceNodeCatalog
 extends Resource